restructure settings page
