+++ clients/web/src/components/settings/SettingsModal.tsx
@@ -1,15 +1,8 @@
 import { useState } from "react";
-import { X } from "lucide-react";
 import { useUser } from "@clerk/clerk-react";
 import { Dialog as DialogPrimitive } from "radix-ui";
-import { SettingsNav } from "./SettingsNav";
-import { ProfileTab } from "./ProfileTab";
-import { MembersTab } from "./MembersTab";
-import { MemberDetailPanel } from "./MemberDetailPanel";
-import type { Member } from "./MembersTab";
-import type { SettingsTab } from "./SettingsNav";
-import { DialogTitle } from "@/components/ui/dialog";
-import { cn } from "@/lib/utils";
+import { SettingsNav, type SettingsTab } from "./SettingsNav";
+import { SettingsContentPanel } from "./SettingsContentPanel";
 
 type SettingsModalProps = {
   open: boolean;
@@ -24,8 +17,6 @@ const tabTitles: Record<SettingsTab, string> = {
 export function SettingsModal({ open, onClose }: SettingsModalProps) {
   const { user } = useUser();
   const [activeTab, setActiveTab] = useState<SettingsTab>("profile");
-  const [selectedMember, setSelectedMember] = useState<Member | null>(null);
-  const [memberNotes, setMemberNotes] = useState<Record<string, string>>({});
 
   const displayName =
     user?.fullName ??
@@ -34,71 +25,17 @@ export function SettingsModal({ open, onClose }: SettingsModalProps) {
   const title =
     activeTab === "profile" ? displayName || "User" : tabTitles[activeTab];
 
-  function handleTabChange(tab: SettingsTab) {
-    setActiveTab(tab);
-    setSelectedMember(null);
-  }
-
-  function handleRoleChange(role: "Admin" | "Member") {
-    if (!selectedMember) return;
-    setSelectedMember((prev) => (prev ? { ...prev, role } : null));
-  }
-
   return (
     <DialogPrimitive.Root open={open} onOpenChange={(o) => !o && onClose()}>
       <DialogPrimitive.Portal>
         <DialogPrimitive.Overlay className="fixed inset-0 z-50 bg-black/40 data-open:animate-in data-open:fade-in-0 data-closed:animate-out data-closed:fade-out-0 duration-300 ease-out" />
         <DialogPrimitive.Content className="fixed left-1/2 top-1/2 z-50 flex h-[90vh] w-[90vw] -translate-x-1/2 -translate-y-1/2 overflow-hidden rounded-xl bg-white shadow-2xl outline-none duration-300 ease-out data-open:animate-in data-open:fade-in-0 data-open:zoom-in-90 data-closed:animate-out data-closed:fade-out-0 data-closed:zoom-out-90">
-          <SettingsNav activeTab={activeTab} onTabChange={handleTabChange} />
-
-          {/* Right content area */}
-          <div className="relative flex-1 overflow-hidden">
-            {/* Close button — always on top */}
-            <button
-              type="button"
-              onClick={onClose}
-              className="absolute right-4 top-4 z-20 rounded-md p-1.5 text-text-subtle hover:bg-bg-selected hover:text-text-default"
-            >
-              <X className="size-4" />
-            </button>
-
-            {/* Main scrollable content */}
-            <div className="h-full overflow-y-auto p-12">
-              <div className="mb-8">
-                <DialogTitle className="text-3xl font-bold text-text-default">
-                  {title}
-                </DialogTitle>
-              </div>
-
-              {activeTab === "profile" && <ProfileTab />}
-              {activeTab === "members" && (
-                <MembersTab onSelectMember={setSelectedMember} />
-              )}
-            </div>
-
-            {/* Member detail panel — slides in over content */}
-            <div
-              className={cn(
-                "absolute inset-0 z-10 bg-white transition-transform duration-300 ease-in-out",
-                selectedMember ? "translate-x-0" : "translate-x-full",
-              )}
-            >
-              {selectedMember && (
-                <MemberDetailPanel
-                  member={selectedMember}
-                  note={memberNotes[selectedMember.id] ?? ""}
-                  onNoteChange={(note) =>
-                    setMemberNotes((prev) => ({
-                      ...prev,
-                      [selectedMember.id]: note,
-                    }))
-                  }
-                  onRoleChange={handleRoleChange}
-                  onBack={() => setSelectedMember(null)}
-                />
-              )}
-            </div>
-          </div>
+          <SettingsNav activeTab={activeTab} onTabChange={setActiveTab} />
+          <SettingsContentPanel
+            activeTab={activeTab}
+            title={title}
+            onClose={onClose}
+          />
         </DialogPrimitive.Content>
       </DialogPrimitive.Portal>
     </DialogPrimitive.Root>
